Primary work on mruby binding
Encouter a lot a of issue since mruby is a full C fledged library,
there is no easy way to handle object method call.
The curremt workaround is to use the mrb_state to fetch the window
that is in the context of mruby.
This is probably inefficient and will lead to nasty code (globals,
for each windows), but this is the easiest way I found for now.
Feel free to suggest an alternative to this.